Revert CGUIGridLayout (move to cgui-gridlayout branch)
This reverts commit dc34672cff71494322e93995de84d089dd79b6ce.

Revert "Implement CGUIGridLayout custom element"

This reverts commit 4f47d471c84f916d659650be66fcefa545ee2997.

Revert "Add map for item lookup"

This reverts commit 5c47f44193f49f840b03f69256cb2668b1bd7975.

Revert "Fix issues with CEGUI testing tab"

This reverts commit 42cdfd4fa3aa881d0ec0616a7d514f1556aa02da.

Revert "Minor optimization in item lookup map usage"

This reverts commit c80bb0c21c06deed5f0ad326b49e3fb0843ed1ed.

Revert "Fix cell item assignment in CGUIGridLayout"

This reverts commit 068a4975661ec5872318218651a5f3b1d06f7f77.

Revert "Add flex column & row support to CGUIGridLayout"

This reverts commit 548e53025d20122eb66987201ded0dca190a18f6.
