+++ Client/core/CChat.h
@@ -13,7 +13,6 @@
 
 #include "CGUI.h"
 #include <core/CCoreInterface.h>
-#include <../Shared/sdk/CColor.h>
 
 class CChatLineSection;
 
@@ -22,6 +21,39 @@ class CChatLineSection;
 #define CHAT_MAX_LINES 100                               // Chatbox maximum chat lines
 #define CHAT_BUFFER 1024                                 // Chatbox buffer size
 #define CHAT_INPUT_HISTORY_LENGTH 128                    // Chatbox input history length
+
+class CColor
+{
+public:
+    CColor() { R = G = B = A = 255; }
+    CColor(unsigned char _R, unsigned char _G, unsigned char _B, unsigned char _A = 255)
+    {
+        R = _R;
+        G = _G;
+        B = _B;
+        A = _A;
+    }
+    CColor(const CColor& other) { *this = other; }
+    CColor(unsigned long ulColor) { *this = ulColor; }
+    CColor& operator=(const CColor& color)
+    {
+        R = color.R;
+        G = color.G;
+        B = color.B;
+        A = color.A;
+        return *this;
+    }
+    CColor& operator=(unsigned long ulColor)
+    {
+        R = (ulColor >> 16) & 0xFF;
+        G = (ulColor >> 8) & 0xFF;
+        B = (ulColor)&0xFF;
+        return *this;
+    }
+    bool operator==(const CColor& other) const { return R == other.R && G == other.G && B == other.B && A == other.A; }
+
+    unsigned char R, G, B, A;
+};
 
 class CChatLineSection
 {
