+++ Client/gui/StdInc.h
@@ -35,4 +35,3 @@
 #include "CGUIWebBrowser_Impl.h"
 #include "CGUIWindow_Impl.h"
 #include "CGUIComboBox_Impl.h"
-#include "CGUIGridLayout_Impl.h"
