+++ Client/sdk/gui/CGUIElement.h
@@ -41,8 +41,7 @@ enum eCGUIType
     CGUI_SCROLLPANE,
     CGUI_SCROLLBAR,
     CGUI_COMBOBOX,
-    CGUI_WEBBROWSER,
-    CGUI_GRIDLAYOUT
+    CGUI_WEBBROWSER
 };
 
 class CGUIElement
